Add support for experimental Esqueleto API, supporting joins
This updates the library to a newer version of Esqueleto which exports
an experimental API, which brings additional safety guarantees in the
type system that help to prevent runtime errors.

Using the experimental API, this adds support for streaming results from
joins, using the new function `streamFrom`, which is more general than
`streamEntities`. A test demonstrates how the function can be used to
stream from an inner join.

This commit contains an orphan instance, which is put under discussion
in github.com/bitemyapp/esqueleto/issues/267 - the instance is necessary
to use `streamFrom` for joins at the moment, but should really be
upstreamed.